Handle all characters in the initial input.
At the moment it's possible to pass extra numbers
as command-line arguments, and they end up on the
stack when the app starts. But it should be possible
to run hc like dc when there is little else needed:

$ hc 123 456 + q

...should return the result right away.

Note that dc uses arguments as a file to read from
but I always found that a bit silly.